feat: multiple selection in chat list
To use it, simply hold Shift or Ctrl and click on the chat list items.
Keyboard is also supported.
Accessibility is taken into account
(`aria-selected`, `aria-multiselectable`).

This has minimal effect on non-multiselect behavior,
i.e. if you don't hold Shift or Ctrl, you won't see a big difference.
The most significant change in this regard is that
if you right-click a chat that is not the active chat,
it will get highlighted, and the active chat will get un-highlighted,
until the context menu gets closed.

Note that the active chat and the selected chats have the same look
(i.e. darker background).
I have tried to minimize confusion between them.
But in the end we might need to style them differently.
Although at the current state it also works quite well IMO.

Summary of the changes:
- Make `ChatListContextMenu` support multiple items
  (without changing behavior for single selection).
  The behavior if things such as "pin / unpin", "mute / unmute"
  is the same as in Android.
  The "Delete Chat" dialog also supports multiple items,
  and displays the to be deleted chat names.
- Rename `selectedChatId` to `activeChatId` in `ChatList.tsx`
- Introduce the `useMultiselect` hook. It basically takes `click`
  and `focus` events, and outputs the set of selected items.
  It basically depends on the existing `useRovingTabindex` hook
  for arrow key focus management.
  I have tried searching for existing such libraries,
  but didn't find one that's sufficiently generic and simple.
- Introduce `useChatListMultiselect` hook that wraps
  `useMultiselect`, and more concretely implements multiselect
  specifically for chat list. For example,
  resets selection when the active chat changes.
- Pass down `onClick` and `onFocus` handlers
  to chat list item components.
- Adjust `onContextMenu` to handle multiselect.

The more useful feature would be to multiselect messages,
but I decided to start with a simpler case,
as I once did with `useRovingTabindex`.